delete user confirmation, ui changes

+++ src/containers/home.js
@@ -4,7 +4,6 @@ import { Button } from "react-bootstrap";
 import Container from "react-bootstrap/Container";
 import axios from "axios";
 import "./home.css";
-import { MDBBtn } from "mdb-react-ui-kit";
 
 export default function Home() {
   const [state, setState] = useState([]);
@@ -75,14 +74,21 @@ export default function Home() {
   };
 
   const deleteUser = (username) => {
-    const users = JSON.parse(localStorage.getItem("users"));
-    const getUsers = users.filter((user) => user.Username !== username);
-    localStorage.clear();
-    localStorage.setItem("users", JSON.stringify(getUsers));
-    setState(getUsers);
-    setEdituser();
-    setSearchuser();
-    alert(`User  with username: ${username} has been removed`);
+    const confirm = window.confirm(
+      "Are you sure you wish to delete this item?"
+    );
+    if (confirm) {
+      const users = JSON.parse(localStorage.getItem("users"));
+      const getUsers = users.filter((user) => user.Username !== username);
+      localStorage.clear();
+      localStorage.setItem("users", JSON.stringify(getUsers));
+      setState(getUsers);
+      setEdituser();
+      setSearchuser();
+      alert(`User  with username: ${username} has been removed`);
+    } else {
+      alert(`User with username: ${username} is not removed`);
+    }
   };
 
   const editUser = (username) => {
@@ -136,14 +142,148 @@ export default function Home() {
 
   return (
     <div>
+      <div className="menubar">
+        <Form.Group controlId="form.user">
+          <p></p>
+          <Form.Control
+            type="text"
+            placeholder="user detail"
+            value={user.userinfo}
+            onChange={(e) => setUserinfo(e.target.value.trimEnd())}
+          />
+        </Form.Group>
+        &nbsp;&nbsp;
+        <Button
+          className="pull"
+          variant="primary"
+          onClick={() => filterTable(userinfo)}
+        >
+          Search
+        </Button>{" "}
+        &nbsp;&nbsp;
+        <p></p>
+ 
+        <button className="register" onClick={() => formDisplay()}>
+          Enter new user
+        </button>
+      </div>
+        <modal>
+          <div id="form" style={{ display: "none" }}>
+            <Container>
+              <h3> Register New User</h3>
+              <Form>
+                <button class="btn-close" onClick={() => formDisplay()}>X</button>
+                <Form.Group controlId="form.Gender">
+                  <Form.Label>Gender: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter gender"
+                    value={user.Gender}
+                    onChange={(e) => {
+                      setUser({ ...user, Gender: e.target.value });
+                    }}
+                    autoComplete="off"
+                  />
+                </Form.Group>
+                <p></p>
+                <Form.Group controlId="form.Name">
+                  <Form.Label>Name: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter full name"
+                    value={user.name}
+                    onChange={(e) => {
+                      setUser({ ...user, Name: e.target.value });
+                    }}
+                    autoComplete="off"
+                  />
+                </Form.Group>
+                <p></p>
+                <Form.Group controlId="form.Email">
+                  <Form.Label>Email: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    autocomplete="false"
+                    placeholder="Enter email"
+                    value={user.email}
+                    onChange={(e) => {
+                      setUser({ ...user, Email: e.target.value });
+                    }}
+                  />
+                </Form.Group>
+                <p></p>
+                <Form.Group controlId="form.Username">
+                  <Form.Label>Username: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter username"
+                    value={user.username}
+                    onChange={(e) => {
+                      setUser({ ...user, Username: e.target.value });
+                    }}
+                    autoComplete="off"
+                  />
+                </Form.Group>
+                <p></p>
+                <Form.Group controlId="form.Password">
+                  <Form.Label>Password: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter password"
+                    value={user.password}
+                    onChange={(e) => {
+                      setUser({ ...user, Password: e.target.value });
+                    }}
+                    autoComplete="off"
+                  />
+                </Form.Group>
+                <p></p>
+                <Form.Group controlId="form.DOB">
+                  <Form.Label>DOB: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter DOB"
+                    value={user.dob}
+                    onChange={(e) => {
+                      setUser({ ...user, DOB: e.target.value });
+                    }}
+                    autoComplete="off"
+                  />
+                </Form.Group>
+                <p></p>
+                <Form.Group controlId="form.Phone">
+                  <Form.Label>Phone number: </Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter phone number"
+                    value={user.phone}
+                    onChange={(e) =>
+                      setUser({ ...user, Phone: e.target.value })
+                    }
+                    autoComplete="phone-number"
+                  />
+                </Form.Group>
+                <p></p>
+                <Button
+                  onClick={() => submitUser(user)}
+                  variant="primary"
+                  type="reset"
+                >
+                  Submit
+                </Button>
+              </Form>
+            </Container>
+          </div>
+        </modal>
       {edituser ? (
         <div>
           <Container>
             <h3> Edit User</h3>
             <Form>
-              <MDBBtn className="btn-close" color="none" aria-label="Close">
-                X
-              </MDBBtn>
+              <button class="btn-close" onClick={() => formDisplay()}>
+                {" "}
+                X{" "}
+              </button>
               <Form.Group controlId="form.Gender">
                 <Form.Label>Gender: </Form.Label>
                 <Form.Control
@@ -238,9 +378,11 @@ export default function Home() {
           </Container>
         </div>
       ) : null}
-      <h2>Existing Users's Details:</h2>
+      <p></p>
+      <p></p>
+      <h2>Users Details:</h2>
       <table class="styled-table">
-        <thead>
+        <thead> 
           <tr>
             <th>Gender</th>
             <th>Name</th>
@@ -285,133 +427,6 @@ export default function Home() {
           })}
         </tbody>
       </table>
-      <p></p>
-      <button onClick={() => formDisplay()}>Enter new user</button>
-      <modal>
-        <div id="form" style={{ display: "none" }}>
-          <Container>
-            <h3> Register New User</h3>
-            <Form>
-              <MDBBtn className="btn-close" color="none" aria-label="Close">
-                X
-              </MDBBtn>
-              <Form.Group controlId="form.Gender">
-                <Form.Label>Gender: </Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter gender"
-                  value={user.Gender}
-                  onChange={(e) => {
-                    setUser({ ...user, Gender: e.target.value });
-                  }}
-                  autoComplete="off"
-                />
-              </Form.Group>
-              <p></p>
-              <Form.Group controlId="form.Name">
-                <Form.Label>Name: </Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter full name"
-                  value={user.name}
-                  onChange={(e) => {
-                    setUser({ ...user, Name: e.target.value });
-                  }}
-                  autoComplete="off"
-                />
-              </Form.Group>
-              <p></p>
-              <Form.Group controlId="form.Email">
-                <Form.Label>Email: </Form.Label>
-                <Form.Control
-                  type="text"
-                  autocomplete="false"
-                  placeholder="Enter email"
-                  value={user.email}
-                  onChange={(e) => {
-                    setUser({ ...user, Email: e.target.value });
-                  }}
-                />
-              </Form.Group>
-              <p></p>
-              <Form.Group controlId="form.Username">
-                <Form.Label>Username: </Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter username"
-                  value={user.username}
-                  onChange={(e) => {
-                    setUser({ ...user, Username: e.target.value });
-                  }}
-                  autoComplete="off"
-                />
-              </Form.Group>
-              <p></p>
-              <Form.Group controlId="form.Password">
-                <Form.Label>Password: </Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter password"
-                  value={user.password}
-                  onChange={(e) => {
-                    setUser({ ...user, Password: e.target.value });
-                  }}
-                  autoComplete="off"
-                />
-              </Form.Group>
-              <p></p>
-              <Form.Group controlId="form.DOB">
-                <Form.Label>DOB: </Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter DOB"
-                  value={user.dob}
-                  onChange={(e) => {
-                    setUser({ ...user, DOB: e.target.value });
-                  }}
-                  autoComplete="off"
-                />
-              </Form.Group>
-              <p></p>
-              <Form.Group controlId="form.Phone">
-                <Form.Label>Phone number: </Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter phone number"
-                  value={user.phone}
-                  onChange={(e) => setUser({ ...user, Phone: e.target.value })}
-                  autoComplete="phone-number"
-                />
-              </Form.Group>
-              <p></p>
-              <Button
-                onClick={() => submitUser(user)}
-                variant="primary"
-                type="reset"
-              >
-                Submit
-              </Button>
-            </Form>
-          </Container>
-        </div>
-      </modal>
-      <p></p>
-      <h2> Search a user:</h2>
-      <Form.Group controlId="form.user">
-        <Form.Label>User: </Form.Label>
-        <Form.Control
-          type="text"
-          placeholder="user detail"
-          value={user.userinfo}
-          onChange={(e) => setUserinfo(e.target.value.trimEnd())}
-        />
-      </Form.Group>
-      <p></p>
-      <Button variant="primary" onClick={() => filterTable(userinfo)}>
-        Search
-      </Button>
-      &nbsp;&nbsp;
-      <p></p>
-    </div>
+      </div>
   );
 }
